[fix]workとdropzoneコンポーネントの修正
