+++ src/components/Elements/FormWorkData.jsx
@@ -1,22 +1,22 @@
-import React ,{useState} from 'react';
-import {postWorksDataRequest} from '../../actions/Works/ActionCreator';
-import {useDispatch } from 'react-redux';
+import React, { useState } from 'react';
+import { postWorksDataRequest } from '../../actions/Works/ActionCreator';
+import { useDispatch } from 'react-redux';
 import '../../styles/Elements/FormWorkData.css'
 
 
 const FormWorkData = () => {
   const dispatch = useDispatch();
   const [workPostTitle, setworkPostTitle] = useState('');
-  const onWorkPostTitleChange = (e) =>{
+  const onWorkPostTitleChange = (e) => {
     setworkPostTitle(e.currentTarget.value);
   }
   const [workPostComment, setworkPostComment] = useState('');
-  const onWorkPostComment = (e) =>{
+  const onWorkPostComment = (e) => {
     setworkPostComment(e.currentTarget.value);
   }
-  const handleClickSubmitt = function(e){
+  const handleClickSubmitt = function (e) {
     e.preventDefault();
-    if(workPostTitle==''|| workPostComment==''){
+    if (workPostTitle === '' || workPostComment === '') {
       return;
     }
     const worksData = [workPostTitle, "file_url", workPostComment];
@@ -29,7 +29,7 @@ const FormWorkData = () => {
 
     <div id="formWorkData">
       <div className="container">
-          <form>
+        <form>
           <div id="wpt">
             <div>
               <input id="workPostTitle" type="text" value={workPostTitle} onChange={onWorkPostTitleChange} placeholder='作品名'></input>
@@ -37,12 +37,12 @@ const FormWorkData = () => {
           </div>
           <div id="wpc">
             <div>
-              <textarea id = "workPostComment" value={workPostComment} onChange = {onWorkPostComment} placeholder='投稿内容'></textarea>
+              <textarea id="workPostComment" value={workPostComment} onChange={onWorkPostComment} placeholder='投稿内容'></textarea>
             </div>
           </div>
-          <input id = 'submit' type="submit" value="投稿する" onClick={ handleClickSubmitt}/>    
-          </form>
-          
+          <input id='submit' type="submit" value="投稿する" onClick={handleClickSubmitt} />
+        </form>
+
       </div>
     </div>
   )
